Add new function setValesFromSegmentRef().
This is like the existing setValesFromSegment() except it takes a reference to the Segment rather than a Segment index.  This is possible to call when the calling code has a reference to the Segment.  Converting a Segment reference to a Segment index to call setValesFromSegment() would be slow and complex, and the very first thing .setValesFromSegment() does is convert the Segment index to a Segment reference,